% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_data}
\alias{sim_data}
\title{simulate peptide level data}
\usage{
sim_data(
  Nprot = 1000,
  probability_of_success = 0.3,
  fc = c(D = -2, N = 0, U = 2),
  prop = c(D = 10, N = 80, U = 10),
  mean_prot = 20,
  sd = 1.2
)
}
\arguments{
\item{Nprot}{number of porteins}
}
\description{
simulate peptide level data
}
\examples{

res <- sim_data()
View(res)
}
